requirements.txt
and==66.0.2
asgiref==3.7.2
astropy==5.3.1
contourpy==1.1.0
cycler==0.11.0
dj-database-url==2.0.0
Django==4.2.3
django-cors-headers==4.2.0
django-heroku==0.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
fonttools==4.42.0
gunicorn==21.2.0
h5py==3.9.0
imageio==2.31.1
kiwisolver==1.4.4
lazy_loader==0.3
llvmlite==0.40.1
markdown-it-py==3.0.0
matplotlib==3.7.2
mdurl==0.1.2
networkx==3.1
numba==0.57.1
numpy==1.24.4
on==0.0.4
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
psycopg2==2.9.6
psycopg2-binary==2.9.9
pyerfa==2.0.0.3
Pygments==2.15.1
PyJWT==2.7.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.1
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0.1
rich==13.5.2
scikit-image==0.21.0
scipy==1.11.1
six==1.16.0
sqlparse==0.4.4
terminal==0.4.0
tifffile==2023.7.18
typing_extensions==4.7.1
tzdata==2023.3
whitenoise==6.5.0
your==0.6.7
